BaseGeometricShape.cs

+++ BaseGeometricShape.cs
@@ -10,9 +10,9 @@ namespace PolymorphismProject
 {
     public class BaseGeometricShape // Base class, BaseGeometricShape
     {
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public virtual int CalculateArea() // Virtual method, CalculateArea
+        public double Width { get; set; }
+        public double Height { get; set; }
+        public virtual double CalculateArea() // Virtual method, CalculateArea
         { 
             return Width * Height;
         }
@@ -28,7 +28,7 @@ namespace PolymorphismProject
 
     public class RighTriangle : BaseGeometricShape // RightTriangle inherited from BaseGeometricShape
     {
-        public override int CalculateArea() // Override method
+        public override double CalculateArea() // Override method
         {
             return base.CalculateArea() / 2;
         }
